fix: keep release retry helper on approved path
