+++ scripts/verify-preview.mjs
@@ -1,8 +1,10 @@
 import { createHash } from "node:crypto";
 import { mkdir, readFile, rename, writeFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
-import { expectedPreviewMediaType } from "./preview-media-types.mjs";
-import { verifyWithTargetedRetries } from "./preview-targeted-retry.mjs";
+import {
+  expectedPreviewMediaType,
+  verifyWithTargetedRetries,
+} from "./preview-media-types.mjs";
 import { assertRouteBuildContract } from "./content-route-build-contract.mjs";
 
 const argumentsList = process.argv.slice(2);
